Updates to network documentation for new features
This change contains fixes to the network documentation as well as
additions of the creation of floating IP and provider networks.
The bonding information was also significantly updated. All references
to --plan overcloud were replaced with --templates. A bug in the docs
referenced an incorrect way to define the NeutronExternalNetworkBridge,
this is fixed now.

Change-Id: Ia2d609c8386f8cf90edb1a0ff57691df88b83aab