Document using latest TripleO related packages
tripleo-ci builds packages on the fly for all the TripleO related
packages. In order to replicate this behavior in our documentation, such
that we have reasonable confidence we're documenting something that has
passed CI, document installing the latest TripleO related packages from
the current Delorean instance.

This also updates the Delorean CI url to use the current-tripleo symlink
which is equivalent to what tripleo-ci is testing.

Change-Id: I12496d4ef9773c103dc8d583b47a863d938656e7