Remove Tuskar from the CLI usage
The command line workflows with Tuskar have proven to be more limited than
using Heat directly. This change removes Tuskar from the CLI documentation and
updates all the examples.

Essentially this means removing all uses of --plan in the deploy command and
all other 'openstack management ...' commands.

Tuskar is still referenced in a number of places and can't be fully removed
until there is an alternative for the Tuskar UI.

Change-Id: I2d4aa114156c5bafb003e6a67862ed6d1449fd98