Document tripleo.sh
Add documentation on how to use tripleo.sh to tripleo-docs. I think just
an overview will be sufficient. I don't feel like we need to go
overboard here, the script is meant to be developer automation, and self
documenting since it's simple bash. No reason to reproduce the same
commands as documention itself.

Change-Id: I4cbd57a32575e1a442c56ebf825e58cf035e19db