Split node configuration and customization docs
There are two distinct interfaces currently described in the
extra_config docs:

1. ExtraConfig *parameters*, which influence the configuration of
services which are already integrated via the default templates

2. ExtraConfig *resources* which are designed to enable node
customization and third-party integration.

So, for the avoidance of confusion, don't conflate these interfaces
in the docs and split them into two pages.

Also, the "customization" part is due for update with several new
interfaces, so the paramters section could easily get missed
at the end of all the resources orientated description.

Change-Id: I64d67f6526c505185dfe663b087b0f0576b09750