Upgrade documentation
This is a first pass at adding upgrades docs which we don't
currently have. For now trying to cover version specific caveats
with a single upgrades guide. We could split this out if it is
preferable. For example, for Liberty-->Mitaka there is an extra
step required after upgrade-init to deliver AODH at
https://review.openstack.org/#/c/320150

Co-Authored-By: Marios Andreou <marios@redhat.com>
Co-Authored-By: Michele Baldessari <michele@acksyn.com>
Change-Id: I4af214d1dfc595ea61be40efbc0a04e44225e8b1