Add Liberty and Mitaka admonitions and use them
Many of our stable branch steps are actually specific to one branch,
so instead of having to remember to note that in the text of those
sections, let's just have an admonition that makes the version
explicit.

This also corrects the repo setup steps to separate the Liberty and
Mitaka repos, and moves the branch options to their own section in
the environment selector.

I did not touch the image building repo setup yet because I plan
to change that in a follow-up commit anyway so it matches what we
actually do in tripleo.sh/CI.

Change-Id: I871651f0dd782ad51127d5ef0b01be3806272820