Group advanced topics by type
We're getting to have a lot of "advanced" docs, and most of those
fit into one of a few categories.  This change separates the
advanced topics into one of the following categories:

-Feature Config
-Baremetal Node Config
-Backend Config
-Custom Config
-Developer Docs

Single image building is moved to post deployment since it's not
a normal part of the initial deployment.

Change-Id: I14e31469a533efbfe3c18c4bd11da8b3f8f455e6