Switch docs to use yaml with overcloud image build
The tripleo-build-images command is a much better implementation of
the image build driver. It uses yaml files to define the images in
a declartive manner, instead of imperatively in code. This patch
does so for the docs. It outlines how to use 'overcloud image build'
with that now calls tripleo-build-images.

Change-Id: I13adeb126d94dc965f58ac39cd6385872405b463