Use centos.org URLs instead of redirecting
Some of the repo files are now available on their new centos.org home
rather than rdoproject.org. This causes problems with some curl calls,
since it doesn't follow redirects by default.

In this patch those URLs which are currently redirected to centos.org
are replaced with the new URLs. Also, in places where curl calls are
executed by the user, the -L option is added to the curl call.

Change-Id: I18b46b0d2cf95b16e5dcb2a090433f6a5bd36cde