Fix environment-specific limiting for upgrade docs
We now properly tag the repository switching admonitions so that they
show up correctly with regards to the state of "Limit Environment
Specific Content" checkboxes for the "Upgrade Version" category.

This means we also have to add Kilo to Liberty upgrade tag so that
Liberty repos are properly tagged and hideable, even though Kilo to
Liberty upgrades aren't really supported. This peculiar situation will
resolve itself once we remove Liberty-specific steps from the
documentation.

Also a minor tweak is removing a TODO comment from the docs. We'll keep
tweaking the docs as we improve he upgrades experience in that regard.

Change-Id: I1f2c85d89016ce30f392fc05aa35d6796588d748