Remove usage of --compute-scale from the scale roles docs
This CLI arg was deprecated in Newton in favour of using the param in a
environment file. The CLI argument is still present but is no longer the
recommended approach and will be removed in the future. It was deprecated as
it maps directly to a Heat parameter and it makes more sense to store it
in an environment file with the other Heat parameters in the deployment.

Change-Id: Ifb212c7ecaa145726b3c324f7064e62e10d006e8