Remove usage of --compute-scale from the baremetal overcloud docs
This CLI arg was deprecated in Newton in favour of using the param in a
environment file. The CLI argument is still present but is no longer the
recommended approach and will be removed in the future. It was deprecated as
it maps directly to a Heat parameter and this it makes more sense to store it
in an environment file with the other Heat parmeters in the deployment.

Change-Id: I4b751cc99a8978a1bcfa83f7bb3f3baeb764c682