Remove usage of --control-scale from the HA docs
This CLI arg was deprecated in Newton in favour of using the param in a
environment file. The CLI argument is still present but is no longer the
recommended approach and will be removed in the future. It was deprecated as
it maps directly to a Heat parameter and this it makes more sense to store it
in an environment file with the other Heat parmeters in the deployment.

Change-Id: Ifc99b2390b29fa3729cfa04d4178ff236a434246