Uses Ceph/Hammer repos for Liberty and Mitaka only
The use of Ceph/Hammer is not for all stable branches but only for
the Liberty and Mitaka stable releases; Newton will use Ceph/Jewel
instead. This change moves the Ceph repo admonitions within the
relevant stable branch.

Also suggests upgrade of 'cephstorage' nodes before 'compute' nodes
so that Ceph cluster is upgraded to the newest release before the
OpenStack upgrade is started.

Change-Id: Ia12e4f7d6fa7b8ecff7dcec6b5cd85804a049070