Also stop openvswitch on undercloud update
CentOS 7.3 is going to include a new version of openvswitch, and
when we update to it the undercloud ip configuration gets broken
unless we first stop openvswitch.  Since we already stop all the
openstack services for similar reasons, we can just add openvswitch
to the list of things to stop.

Change-Id: I653a6b19d1417ab30d5cc4004638f3b942e9304c