Clean up admonition problems
There have been a couple of admonitions added recently that were
not completely wired into the admonition system, which makes them
essentially glorified note sections instead of dynamic admonitions.
This patch finishes the configuration of the admonitions.

It also does some re-organizing of the existing admonitions.  The
SSL admonitions are dropped in favor of a generic "Optional"
admonition that can be used for these one-off optional features.

Change-Id: I9fd5c3a7f50e4916e45ad16a6e9aaaffd83affe8