Enable warning-is-error for docs job
As of I319e0d4e4dd03b7f769d9d38ad2829d6cc9a679a, the warnings in the
tripleo-docs build should be fixed and it would be good to keep it
that way.

This was added in sphinx 1.5, which is now in global-requirements.
We should add a requirements sync job to this repo so we don't get
out of date again.

Note that since this was originally proposed, some new warnings
have been added to the repo.  These are also fixed in this commit.

Change-Id: Ibc673109878d5edc9a2772e7e0650a1ed31a6cc7