Add docs around updating undercloud ssh key
The ssh key for the undercloud user is used in provisioning new nodes so
special care must be taken when updating the key to ensure the operator
is not locked out of new and old nodes. This change provides a bit of
documentation around the process for rotating the ssh key for the user
on the undercloud.

Change-Id: Ia52b775fe6f88f17961e812e03a4f57a93c77f00