Re-enable admonition selector for openstackdocstheme
In the transition to openstackdocstheme we lost the admonition
selector.  This is because the new theme doesn't provide the same
extension points as the old one, so we were trying to inject the
selector into a non-existent block.

This change moves it to a block that does exist in the theme and
tweaks the styling a bit so it fits better in its new place.  Note
that this does result in it becoming part of the page header area
as opposed to the sidebar.  There doesn't appear to be any block in
the sidebar that we can use for this purpose.  I think we should go
ahead and fix the selector this way, and if we decide we can't
stand the new location we can work with the docs team to get a new
block in the sidebar that we can use.

Change-Id: I10d8f701a108d20f77cf1ab7538b0eb7eb313601
Closes-Bug: 1702523