Update Pike storage documentation for ceph-ansible
The user is required to install ceph-ansible in pike.
The package can not be made a dependency of tripleoclient [1]
Updated install, update/upgrade, and backup and restore

The installation section was updated to cover deployment
of Ceph via TripleO with either puppet-ceph (Ceph not in
containers) or ceph-ansible (Ceph in containers).

[1] http://eavesdrop.openstack.org/meetings/rdo_meeting___2017_06_21/2017/rdo_meeting___2017_06_21.2017-06-21-15.01.log.html#l-29

Co-Authored-By: John Fulton <fulton@redhat.com>
Change-Id: I8e9def9577dc84b63895e10924a01a4949d02aef