Don't use deprecated/removed cli args in net-iso docs
The neutron-* cli args have been deprecated for multiple releases
and are completely removed in Pike.  However, the documentation has
still been referencing them all this time.  This change replaces all
of those instances with examples of the correct parameter values in
network-environment.yaml.

There's no need for a stable version of the commands since the
cli args were already deprecated in Newton, which is the oldest
supported release at this time.

Change-Id: I83fa7d62e573b73450e897e4eb4da5ebde5defa7
Closes-Bug: 1703932