Document the need to clean the Ceph disks when redeploying
When repeating an overcloud deployment which reuses the same nodes
or disks of a previous deplyment, it is necessary to clean the
Ceph OSD disks for the activation to complete. Both puppet-ceph and
ceph-ansible have the same requirement.

This change documents use of Ironic automated_clean functionality
to automate the disks cleanup.

Change-Id: I08ba35c59d3a6aaed59a1526757a4fa513acb190
Related-Bug: #1714242