Clarify predictable ip behavior for deleted nodes
Entries should never be removed from the predictable ip lists, even
if a node is deleted.  This is because the index into the list is
based on the Heat resource index, which doesn't change when a node
is deleted.  This has come up a number of times in real-world
deployments so we should make this behavior clear in the docs.

Change-Id: I455a1e15bfc3cef07f6fd3c8462428985f12c0da