docs: installation: remove unnecessary blockquotes
Due to missing spaces, some of the content in the installation
guide is displayed as blockquote without being really quoted
content. This change adds the needed spaces to remove the
generated HTML blockquotes.

The repositories.txt file gets included at different locations
and that file contained the formatting spaces which were needed
from the location which includes that file. It's most of the time
better to do the indentation at the ".. include" level and leave
the file which gets included *without* indentation. Because here,
the indentation in the included file worked for some includes but
not for all, because the content got used at different indentation
levels. That made it necessary to touch the includes in
basic_deployment and post_deployment too.

Follow up patches will deal with the other install guide sections
more in details.

Change-Id: Ie2e7fea6d92f7742d23e22452e394b8ff40975eb