Document simpler undercloud major upgrade
The undercloud update docs were changed to reflect the new process
that started in Pike, but the major upgrade docs were not.  While the
old process will work for now, there are plans to make changes that
will rely on the new process so we need to make sure to document it.

See 1e731bd20187700ea9f64ca9e346cf65ba4c8d23

bp undercloud-upgrade

Change-Id: I8704b197471f5e53377a7a5f18cdcc43563c54cb