Document use of undercloud_hostname
undercloud_hostname is a less error-prone method of configuring all
the undercloud hostname settings correctly.  We probably want to
make it a mandatory option and avoid hostname-related bugs, but for
the moment let's at least document it as the preferred way to handle
hostname configuration.

Note that this option has existed since Mitaka, so it applies to
all of our currently supported branches.

Change-Id: I2506a44abac999d4b0450c402fe40cab89410cea