Unify undercloud upgrade steps
The steps needed to upgrade the undercloud were
specified in two different files. This commit
updates the steps to cover O to P from commit
496223 in the 'Undercloud Installation' section
and reuses that doc file for the 'Upgrading to
a Next Major Release'.

As an addition, a new class 'otop' has been
added to include specific documentation sections
when upgrading from Ocata to Pike.

Change-Id: I9097ca11a95f94c5eaf580fb8a82b3173d606a5d