Cleanup remaining pointers to storage-environment.yaml
We used to have references to the storage-environment.yaml Heat
environment file when documenting the deployment of Ceph.

From the newton release instead we suggest use of puppet-ceph.yaml
and, from the pike release, ceph-ansible.yaml. These enable
respectively enable use of puppet-ceph or ceph-ansible for the
deployment of Ceph.

This submission removes the remaining references to the
storage-environment.yaml file, hopefully avoiding confusion in
the future by making the choice more explicit.

Change-Id: I4d936ac3988d6fc7d3bb0b18535045137a0d11f7