Update docs to reflect updates to os-net-config mechanism
The deprecated os-net-config element is being removed, so older
templates that use the os-apply-config header will no longer
work. This change updates the documentation with the changes
required to update the templates. It also removes the
documentation about updating the templates for static IPs,
since these changes have been required for a very long time.
A few IP addresses and VLANs were also corrected for clarity.

Change-Id: I68531a6e412afc0db3abfc4984404b8a9f963b55