Update the driver documentation for the undercloud
This change updates the wording to use "hardware types", which is our
primary notion right now. A difference between hardware types and classic
drivers is mentioned, and a few helpful links are added.

Add the full list of (techinically) supported drivers, and link to their
documentation pages. Stop "recommeding" vendor drivers, as they often have
problems that the IPMI driver does not have, and make it clear that these
drivers are not supported by the TripleO team (incl. its ironic part).

Change-Id: I8b2b307c9dd0d3d53e2ab791184128628c72c43c
Related-Bug: #1690185