Node specific hiera data doc should mention inspection_extras
For node specific hiera data to work the machine unique UUID must
be in the Ironic database. However it's possible it will be missing
unless the undercloud is configured to collect it during
introspection. This change adds a sentence pointing out that this
could happen so the reader knows what to do if the machine unique
UUID is missing.

Change-Id: I754845be0a9c64fc5f7e9f6b3ec94d7bdab870b2