Simplify container prepare instructions
Now that the prepare command supports --tag-from-label discovery on
every image, the separate discover command is no longer required.

This change also switches the image namespace to
docker.io/tripleomaster. trunk.registry.rdoproject.org was never
intended to be used for deploying openstack from since it is a piece
of infrastructre used in the RDO image build pipeline.
docker.io/tripleomaster is more appropriate since it contains the
tested images which are also hosted on trunk.registry.rdoproject.org.

Change-Id: I4fd54d4374bde9120aa2c74725bfd2db4b3cf3d0
Closes-Bug: #1741544