Doc config-download with deployed-server
Using config-download with deployed-server requires the use of a
HostnameMap since the configuration generated by Heat needs the hostname
value, and we can't know it until we use ansible to get the value.

A HostnameMap parameter value works around that limitation by explicitly
telling Heat what the hostname value is for each deployed-server.

Change-Id: Ie2f644a7d4b6ab8ec3eaefa138fd029c0a538a9c