undercloud/upgrade: suggest backup before upgrade
Suggest to perform a backup before the undercloud upgrade and validate
that the restore works fine. It's to make sure our operators understand
this procedure is critical.
We'll also add some warning in tripleoclient so our operators are asked
to do a backup before proceeding to the upgrade.

Also removing the note about the fact any package can be upgraded. This
note is not useful as we're moving to containers anyway.

Change-Id: I1f076105a47b70d11311a8c950e0fe5760d6fd5f