Clarify docker vs moby containers runtime/images
For containerized undercloud as a good start, avoid using
references to Docker if only not limited to the docker CLI
commands.

Whenever we refer to *architecture* for containers runtime and
images (CRI & OCI), refer to Moby instead.

This creates less confusion for readers might be expecting more
features of Docker is used, like Swarm as COE, while it is not.

Eventually, this replacement should happen to all places.

Change-Id: Ia731d7a151f0a294297b0528173e36577ad0dfeb