Correcting split-stack document
The deployed-server-environment-output.json format generated
from the value of the DeployedServerEnvironment output of the
overcloud-baremetal stack is not a valid json format. And service
deployment command ultimately fails due to invalid json format.

Changing the output format to json and adding a jq filter to get
json format of output value.

Change-Id: I19ee30f20ce3c02a4c4d0dfb2b522818a15f629f