Add usage examples for standalone
Recently folks have been asking how to launch the VMs in the standalone
and be able to access them. The following patch provides examples on how
to consume the standalone resources and leverage tenant and provider
networks for the VMs and how to use only provider networks for the VMs.

Change-Id: I32b51374376db92ee6854fa7816b59bf12f97fc6