config-download User's Guide
Adds a full User's Guide for config-download, documenting most
everything I could think of that would be useful.

Previously the page was under the "Feature Configuration" section, and
it's now moved to it's own top level section witin the ToC.

There are a few other minor fixes and improvements included.

Change-Id: I9bc2df4f44ac4ccea565d632d0072c2c4d7b3a43