Make every reference to the prepare command use sudo
Even though the installing user is added to the docker group, this
only takes effect after they restart their shell session. So if they
install the undercloud and run prepare in the same session, the
prepare call will fail at the docker operations.

Rather than explain this caveat, document calling prepare with sudo.
This is already done on line #266, but this change ensures every
reference to the command includes sudo.

Change-Id: Ia14b0964dcaefbbf7265b3960315114219fd3c7b
Closes-Bug: #1793375