Remove collab_upgrades from dev docs
This never got written, and the other dev docs should be detailed
enough to expose inner workings of updates/upgrades/ffwd, and reading
them and scanning code for existing services should give enough
insight to be able to contribute code.

If we still wanted to provide a step-by-step tutorial, we can
reintroduce the page later, leaving WIP stubs of pages can
confuse/demotivate the reader, so we shouldn't have them there for
extended periods of time.

Change-Id: I7bf4a57efede706fe5fa639a637298ca884a8783