Don't keep shrinking font when nesting admonitions
In some places we nest admonitions, e.g. a note within a
release-version-specific part of docs. Each admonition has a font-size
of 85% of its parent element. Even for the first nested admonition
this results in a barely readable text.

Fix it so that only the first admonition shrinks font to 85% of
parent, and it doesn't shrink further when nesting.

Change-Id: I91992a29375eea1ca9971d05827a20153b1721a7