Deployed Server - get-occ-config.sh need ordered host list
Make note in documentation that it is critical to list hosts
according to node %index% number when using get-occ-config.sh
to distribute Heat agent config for pre-deployed servers.

If the nodes are not listed in the correct order the Heat
config added will/can be for a different node.

Related: RHBZ#1647363
Related-Bug: #1804642
Change-Id: Ie53dfae7a35b47528aa6d8eca64b1da5b1cedaf7