Add service template sections walkthough (part 1)
This adds the first bit of going through all of the available sections
of the service templates. This will hopefully make it easier for folks
to know what's available and how to use it.

This first part adds the bare minimum you need to do to create a
service, and describes each section.

The rest of the sections will be added in subsequent patches.

Change-Id: I1c8e7f5a8967a68554f28a66d4a151c44d892247
Related-Bug: #1804642