Enhancement to cell v2 doc with split for stein/train
In train we'll add a new tripleoclient command to help with
export stack data which is required for the cell stack input.

Also this change adds a basic and an advanced cell deployment
architecture. Due to release specific features the doc is split
into in different parts (stein,train and cell management)

Depends-On: https://review.opendev.org/672070
Change-Id: Ia14e01f7f8deb55a6ac4f1b30a84352822143c27