Update docs for rerunning undercloud install
The note about not running the undercloud install command if an
overcloud were already deployed was outdated. We now allow undercloud
install reruns when the overcloud is already deployed. In some cases it
is required, such as when adding new ctlplane subnets for spine/leaf
provisioning.

This patch updates the note and adds some points to watch out for when
making changes.

Change-Id: I18aaee1467d0f7b11d3a5a43b3ee0b35e9f90148