Remove references to writeable user and groups for LDAP
Prior to the Ocata release, keystone supported the ability to make
changes to user and group references in LDAP. This functionality was
deprecated and finally removed in Ocata.

We should clarify the tripleo documentation so that it doesn't give
the impression the functionality still exists.

Here is a link to the release notes that advertize the removal:

  https://docs.openstack.org/releasenotes/keystone/ocata.html#relnotes-11-0-0-origin-stable-ocata-other-notes

Change-Id: I2a1d92b6d2aecc11e3e13120204bfc684d8a802a