Document how to use NodeDataLookup generation utility
The updated version of the Ceph documentation warns
about the problems of using name-based methods of
refering to block devices and suggests use of the
by_path method to avoid this problem. It then links
to the updated Node Specific Hieradata document which
contains a description of how to use the utility from
the related bug.

Change-Id: I2131d87c51f142166f341d2a9c209162659fca43