Update the ldap configuration guide to use correct defaults
This change goes hand-in-hand with another patch to puppet-keystone [0]
that removes default values from puppet-keystone. If configuration
values aren't defined in keystone's configuration file, it will use it's
defaults. Having defaults in two places just makes it easier for them to
get out-of-sync, which is the case in 713708.

This updates the docs to reflect the default values in keystone.

[0]  https://review.opendev.org/#/c/713708/

Change-Id: I95c414be4881a230466527e399f2490659a6b61e