Document ceph dashboard Safe mode feature
The Ceph dashboard is deployed using read-only privileges
for admin user, so no ceph cluster changes can be made
(e.g., pool creation).
The purpose of this new paragraph is to document the Safe
mode feature and how to alter the behaviour of the admin
user, providing fully administrative privileges.

Change-Id: Id98f49fd314ae4521fb73c26fae5bfaeea1abc4d
Signed-off-by: Francesco Pantano <fpantano@redhat.com>