Use GlanceImageImportPlugin instead of ExtraConfig overrides for DCN
Following up on change I5c0788bb361323f5a0e69e22d4bd42fa3c1e06ed we shouldn't need
to override the puppet hieradata to configure the list of Glance import plugins.

The parameter in the Heat templates existed before but it wouldn't override the logic
which in the template enabled `image_conversion` plugin even when user would purposely
set it to a different list.

Change-Id: I7fa9034c25887fb96e2d250f459654e6ee798158