Document 'openstack overcloud export ceph' for DCN
This update removes more than it adds becuase the new
'openstack overcloud export ceph' command provided by
the depends-on patch automates the information gathering
necessary to create a Heat environment file with the
CephExternalMultiConfig paramter populated.

This change also gives the deployer the option to
distribute the default openstack cephx keyring or to
create an additional external keyring.

Change-Id: I6d7b568f65315253071985296e0b73c202c3c072
Related-Bug: #1895034