Document the new CephConfigPath option
This change is supposed to document the CephConfigPath option
introduced by the TripleO Ceph client role.
The purpose of this doc change is to make the operators aware
that this option changes in tripleo-heat-templates the default
location of the rendered keyfiles and ceph.conf within the host,
leaving untouched the mountpoint inside the containers.

Change-Id: I2dbedb715097206ed62aae32c83af9fc19932bd1