Update tox.ini for tripleo-docs
Removes the pep8 and bindep tox environments. pep8 did not run
successfully as there is no test-requirements.txt file, and there is
nothing to test anyway. bindep also has nothing to test, so it's also
removed.

The deploy-guide tox environment is added to the default list so that
just running "tox" results in a full build of the documentation
including the user and deploy guide.

Change-Id: I1990556fec47905d2e137b51bcaefd15fe597dea