Replace existing virt-customize usage with tripleo-mount-image
The `tripleo-mount-image` tool was created as an alternative for
kpartx-based customizations which do not support the multiple mount
points of overcloud-hardened-uefi-full.qcow2.

virt-customize sets a machine-id in the image, which results in all
overcloud nodes having the same machine-id [1]. virt-customize can
also be very slow on virtualized underclouds.

This change replaces virt-customize usage with
tripleo-mount-image+chroot as the recommended tool for overcloud image
customizations.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1554546

Change-Id: I3f65479777cc3dfa1ed3c396dbfbe182288ae65b