Update docs for CentOS9
We still suggest installing on CentOS 8 Stream. This wouldn't be
expected to work as we have moved to CentOS Stream 9 (aka. CentOS 9)
for master now. This patch updates our install guide for CentOS 9.

Note that there are still remaining reference to CentOS 8 Stream, but
this change aims to fix what we can fix easily now.

Change-Id: Icb8c584e1a62278ca2d8239a83f603c2e90ec03a